feat: Improve blip archive feature (#62)
This change modifies the archive functionality to be applied
in the blip history. This means we can track when something
is moved to/from archive in the blip history. The old `active`
flag has been removed from the schema.

BREAKING CHANGE: The YAML schema no longer supports the `active`
flag. Instead use the `ARCHIVE` ring type on blip level.